Add SAMRecordWritable and its dependencies.
We need to duplicate much of Picard due to behaviour we need to change
(or its dependencies) being package-private or private and thus not
fixable by simply making a subclass.

The root cause is that the BAMRecord class, an instance of which is
returned by BAMRecordCodec.decode(), requires a valid SAMFileHeader to
be passed to it on creation. This should not actually be necessary, but
that's how it's written. Thus we need to make a custom BAMRecord—
LazyBAMRecord—that doesn't have this limitation.

Unfortunately, BAMRecord is package-private, so we need to copy that
here. And most unfortunately, it depends on some package-private methods
in SAMRecord which modify some private data in SAMRecord, so we need to
pull in SAMRecord as well. These two classes then force us to pull in
all the rest in a bit of a dependency cascade.

SAMRecordWritable notes that we may need to do something similar for
BAMRecordCodec.encode() in case its implementation ever changes to
require a nonnull header (which, again, it should not).