Make SAMFileReader accept a SeekableStream index.
Not just Files. This is needed for an upcoming utility.

Yes, all the other changes are necessary since we need to pull in any
package-private dependencies (of dependencies of dependencies... etc.)
of the classes that we actually needed to modify. Those are:

SAMFileReader
BAMFileReader
CachingBAMFileIndex
DiskBasedBAMFileIndex
AbstractBAMFileIndex

Most notably the last, because that's where the File was actually used.
Making a generic SeekableStream work required adding support for
streaming I/O instead of the memory-mapped kind.

The modifications to existing files are due to having added some of the
classes that were dependencies of the old classes we had copied.