Use stock Picard almost everywhere, update to 1.73
net.sf.samtools.SAMFileReader still can't read a BAM index from a
SeekableStream, which is why cli.plugins.View and util.hadoop.BAMReader
are stuck with the custom Picard. Fortunately they're so isolated from
the rest of the codebase that the rest of the library were able to be
switched.

Apart from imports and the build.xml and Picard .jar file changes, the
only changes here are in LazyBAMRecord, since it was accessing SAMRecord
fields that are private in the stock Picard. Fortunately there's an easy
(though somewhat fragile to changes in upstream behaviour) workaround
using the public API.

Note that the custom Picard hasn't been (and won't be, unless very
necessary) updated, it's still based on 1.56 and older versions plus
some patches of mine. Hence this commit had to include both the 1.73
update as well as switching to use the stock Picard: if the update
hadn't been done, the code wouldn't compile because we depend on
post-1.56 features, and if the switch hadn't been done, the code
wouldn't run because some of the custom code doesn't work with 1.73.
Thankfully the parts that cli.plugins.View uses do work.