CLI: refactor commonalities into CLIMR{,BAM}Plugin
This isn't all-inclusive: some copy-pasted things still remain, such as
the SAM/BAM merging done in both fixmate and sort.

Fixes an error in SummarySort: it used to have an "output-dir" option
when it was supposed to be "output-file".

Now everything is unified to use "output-merged-path", bumping all
affected plugin versions by 0.1, with the exception of the unreleased
fixmate plugin.