Add BCF output support, with AnySAM-like selection
That is, there is only one VCFOutputFormat (and the KeyIgnoring one),
but it behaves for VCF/BCF like (KeyIgnoring)AnySAMOutputFormat does for
SAM/BAM: you can select the format you want and it'll create a
RecordWriter of the appropriate type.

We only output BGZF-compressed BCF, because the spec describes it as the
canonical format: "it's acceptable to not BGZF compress a BCF2 file, but
not all readers may handle this uncompressed encoding".