[BV]CFRecordWriter: allow setting input header
The header of the input records is not necessarily the same as the one
used for output, and the lazy decoding relies specifically on the input
header.

Of course each input record could have a different header, but we only
allow setting one. What should be done for the general case isn't very
obvious...