Use htsjdk machinery to read with intervals significantly faster.
The main idea is to add the chunk intervals for each split into
the FileVirtualSplit class, so that each record reader can use
that information to only read the chunks in the BAM file that it
needs to, rather than read everything and apply filtering afterwards.

This change means that an index is always needed when using
intervals.

BAMRecordReader changes to use an iterator from SamReader, rather
than manage its own iteration (via BAMRecordCodec).

Support for keeping paired reads together in the same split has
been removed (previously set via
'hadoopbam.bam.keep-paired-reads-together'), since it is not
compatible with BAMFileReader managing the chunks to iterate over.
Applications that require this functionality can use the technique
in GATK's BwaSparkEngine#putPairsInSamePartition to achieve the
same result.