Eliminate buffer copy bug in BGZFSplitGuesser.
In the existing BGZFSplitGuesser code, data from the InputStream is copied into
a byte array, which is used as a buffer for finding the next BGZF block start.
Previously, this code assumed that a non-negative return from the read call
meant that the full amount of bytes requested were copied. This does not always
happen, which can break the split guessing code. This patch adds logic to make
sure that the buffer is filled. Additionally, this patch rewrites the BGZF
header guessing code.