Update BGZF block decode logic and refactor BGZFSplitGuesser on BaseSplitGuesser
Stems out of work done in #142 and #149. Refactors BGZFSplitGuesser to use the
logic in the BaseSplitGuesser class. To do this, we need to make
BaseSplitGuesser public. Additionally, we replace the logic that is used in
BaseSplitGuesser to identify the start of a BGZF block with logic that doesn't
require copying bytes from the input stream to a buffer. This improves the
performance of identifying the start of a BGZF block; for whatever reason, that
copy to the buffer/methods on the buffer are expensive.